Refactor to remove aperture_funcs
