+++ photutils/aperture/core.py
@@ -75,6 +75,88 @@ def _make_annulus_path(patch_inner, patch_outer):
     return mpath.Path(verts, codes)
 
 
+def get_phot_extents(data, positions, extents):
+    """
+    Get the photometry extents and check if the apertures is fully out of data.
+
+    Parameters
+    ----------
+    data : array_like
+        The 2-d array on which to perform photometry.
+
+    Returns
+    -------
+    extents : dict
+        The ``extents`` dictionary contains 3 elements:
+
+        * ``'ood_filter'``
+            A boolean array with `True` elements where the aperture is
+            falling out of the data region.
+        * ``'pixel_extent'``
+            x_min, x_max, y_min, y_max : Refined extent of apertures with
+            data coverage.
+        * ``'phot_extent'``
+            x_pmin, x_pmax, y_pmin, y_pmax: Extent centered to the 0, 0
+            positions as required by the `~photutils.geometry` functions.
+    """
+
+    # Check if an aperture is fully out of data
+    ood_filter = np.logical_or(extents[:, 0] >= data.shape[1],
+                               extents[:, 1] <= 0)
+    np.logical_or(ood_filter, extents[:, 2] >= data.shape[0],
+                  out=ood_filter)
+    np.logical_or(ood_filter, extents[:, 3] <= 0, out=ood_filter)
+
+    # TODO check whether it makes sense to have negative pixel
+    # coordinate, one could imagine a stackes image where the reference
+    # was a bit offset from some of the images? Or in those cases just
+    # give Skycoord to the Aperture and it should deal with the
+    # conversion for the actual case?
+    x_min = np.maximum(extents[:, 0], 0)
+    x_max = np.minimum(extents[:, 1], data.shape[1])
+    y_min = np.maximum(extents[:, 2], 0)
+    y_max = np.minimum(extents[:, 3], data.shape[0])
+
+    x_pmin = x_min - positions[:, 0] - 0.5
+    x_pmax = x_max - positions[:, 0] - 0.5
+    y_pmin = y_min - positions[:, 1] - 0.5
+    y_pmax = y_max - positions[:, 1] - 0.5
+
+    # TODO: check whether any pixel is nan in data[y_min[i]:y_max[i],
+    # x_min[i]:x_max[i])), if yes return something valid rather than nan
+
+    pixel_extent = [x_min, x_max, y_min, y_max]
+    phot_extent = [x_pmin, x_pmax, y_pmin, y_pmax]
+
+    return ood_filter, pixel_extent, phot_extent
+
+
+def find_fluxvar(data, fraction, error, flux, imin, imax, jmin, jmax,
+                 pixelwise_error):
+
+    if isinstance(error, u.Quantity):
+        zero_variance = 0 * error.unit**2
+    else:
+        zero_variance = 0
+
+    if pixelwise_error:
+
+        subvariance = error[jmin:jmax, imin:imax] ** 2
+
+        # Make sure variance is > 0
+        fluxvar = np.maximum(np.sum(subvariance * fraction), zero_variance)
+
+    else:
+
+        local_error = error[int((jmin + jmax) / 2 + 0.5),
+                            int((imin + imax) / 2 + 0.5)]
+
+        fluxvar = np.maximum(local_error ** 2 * np.sum(fraction),
+                             zero_variance)
+
+    return fluxvar
+
+
 class _ABCMetaAndInheritDocstrings(InheritDocstrings, abc.ABCMeta):
     pass
