+++ photutils/aperture/ellipse.py
@@ -2,15 +2,16 @@
 from __future__ import (absolute_import, division, print_function,
                         unicode_literals)
 import math
+import warnings
 
 import numpy as np
 from astropy.coordinates import SkyCoord
 import astropy.units as u
 from astropy.wcs.utils import skycoord_to_pixel
+from astropy.utils.exceptions import AstropyUserWarning
 
 from .core import (SkyAperture, PixelAperture, _sanitize_pixel_positions,
-                   _make_annulus_path)
-from .aperture_funcs import do_elliptical_photometry, get_elliptical_fractions
+                   _make_annulus_path, get_phot_extents, find_fluxvar)
 from ..utils.wcs_helpers import (skycoord_to_pixel_scale_angle, assert_angle,
                                  assert_angle_or_pixel)
 
@@ -20,6 +21,153 @@ skycoord_to_pixel_mode = 'all'
 
 __all__ = ['SkyEllipticalAperture', 'EllipticalAperture',
            'SkyEllipticalAnnulus', 'EllipticalAnnulus']
+
+
+def do_elliptical_photometry(data, positions, a, b, theta, error,
+                             pixelwise_error, method, subpixels, a_in=None):
+
+    extents = np.zeros((len(positions), 4), dtype=int)
+
+    # TODO: we can be more efficient in terms of bounding box
+    radius = max(a, b)
+
+    extents[:, 0] = positions[:, 0] - radius + 0.5
+    extents[:, 1] = positions[:, 0] + radius + 1.5
+    extents[:, 2] = positions[:, 1] - radius + 0.5
+    extents[:, 3] = positions[:, 1] + radius + 1.5
+
+    ood_filter, extent, phot_extent = get_phot_extents(data, positions,
+                                                       extents)
+
+    flux = u.Quantity(np.zeros(len(positions), dtype=np.float), unit=data.unit)
+
+    if error is not None:
+        fluxvar = u.Quantity(np.zeros(len(positions), dtype=np.float),
+                             unit=error.unit ** 2)
+
+    # TODO: flag these objects
+    if np.sum(ood_filter):
+        flux[ood_filter] = np.nan
+        warnings.warn("The aperture at position {0} does not have any "
+                      "overlap with the data"
+                      .format(positions[ood_filter]),
+                      AstropyUserWarning)
+        if np.sum(ood_filter) == len(positions):
+            return (flux, )
+
+    x_min, x_max, y_min, y_max = extent
+    x_pmin, x_pmax, y_pmin, y_pmax = phot_extent
+
+    if method == 'center':
+        use_exact = 0
+        subpixels = 1
+    elif method == 'subpixel':
+        use_exact = 0
+    else:
+        use_exact = 1
+        subpixels = 1
+
+    from ..geometry import elliptical_overlap_grid
+
+    for i in range(len(flux)):
+
+        if not np.isnan(flux[i]):
+
+            fraction = elliptical_overlap_grid(x_pmin[i], x_pmax[i],
+                                               y_pmin[i], y_pmax[i],
+                                               x_max[i] - x_min[i],
+                                               y_max[i] - y_min[i],
+                                               a, b, theta, use_exact,
+                                               subpixels)
+
+            if a_in is not None:
+                b_in = a_in * b / a
+                fraction -= elliptical_overlap_grid(x_pmin[i], x_pmax[i],
+                                                    y_pmin[i], y_pmax[i],
+                                                    x_max[i] - x_min[i],
+                                                    y_max[i] - y_min[i],
+                                                    a_in, b_in, theta,
+                                                    use_exact, subpixels)
+
+            flux[i] = np.sum(data[y_min[i]:y_max[i],
+                                  x_min[i]:x_max[i]] * fraction)
+
+            if error is not None:
+                fluxvar[i] = find_fluxvar(data, fraction, error, flux[i],
+                                          x_min[i], x_max[i], y_min[i],
+                                          y_max[i], pixelwise_error)
+
+    if error is None:
+        return (flux, )
+    else:
+        return (flux, np.sqrt(fluxvar))
+
+
+def get_elliptical_fractions(data, positions, a, b, theta,
+                             method, subpixels, a_in=None):
+
+    extents = np.zeros((len(positions), 4), dtype=int)
+
+    # TODO: we can be more efficient in terms of bounding box
+    radius = max(a, b)
+
+    extents[:, 0] = positions[:, 0] - radius + 0.5
+    extents[:, 1] = positions[:, 0] + radius + 1.5
+    extents[:, 2] = positions[:, 1] - radius + 0.5
+    extents[:, 3] = positions[:, 1] + radius + 1.5
+
+    ood_filter, extent, phot_extent = get_phot_extents(data, positions,
+                                                       extents)
+
+    fractions = np.zeros((data.shape[0], data.shape[1], len(positions)),
+                         dtype=np.float)
+
+    # TODO: flag these objects
+    if np.sum(ood_filter):
+        warnings.warn("The aperture at position {0} does not have any "
+                      "overlap with the data"
+                      .format(positions[ood_filter]),
+                      AstropyUserWarning)
+        if np.sum(ood_filter) == len(positions):
+            return np.squeeze(fractions)
+
+    x_min, x_max, y_min, y_max = extent
+    x_pmin, x_pmax, y_pmin, y_pmax = phot_extent
+
+    if method == 'center':
+        use_exact = 0
+        subpixels = 1
+    elif method == 'subpixel':
+        use_exact = 0
+    else:
+        use_exact = 1
+        subpixels = 1
+
+    from ..geometry import elliptical_overlap_grid
+
+    for i in range(len(positions)):
+
+        if ood_filter[i] is not True:
+
+            fractions[y_min[i]: y_max[i], x_min[i]: x_max[i], i] = \
+                elliptical_overlap_grid(x_pmin[i], x_pmax[i],
+                                        y_pmin[i], y_pmax[i],
+                                        x_max[i] - x_min[i],
+                                        y_max[i] - y_min[i],
+                                        a, b, theta, use_exact,
+                                        subpixels)
+
+            if a_in is not None:
+                b_in = a_in * b / a
+                fractions[y_min[i]: y_max[i], x_min[i]: x_max[i], i] -= \
+                    elliptical_overlap_grid(x_pmin[i], x_pmax[i],
+                                            y_pmin[i], y_pmax[i],
+                                            x_max[i] - x_min[i],
+                                            y_max[i] - y_min[i],
+                                            a_in, b_in, theta,
+                                            use_exact, subpixels)
+
+    return np.squeeze(fractions)
 
 
 class SkyEllipticalAperture(SkyAperture):
